+++ module-application/src/main/java/com/zoopi/client/certification/controller/BloodDonationCertController.java
@@ -1,4 +1,4 @@
-package com.zoopi.client.controller.certification;
+package com.zoopi.client.certification.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 
-import com.zoopi.client.api.certification.BloodDonationCertApi;
+import com.zoopi.client.certification.api.BloodDonationCertApi;
 import com.zoopi.ResultResponse;
-import com.zoopi.client.service.certification.BloodDonationCertService;
+import com.zoopi.client.certification.service.BloodDonationCertService;
 
 @RestController
 @RequiredArgsConstructor
