+++ module-application/src/main/java/com/zoopi/client/phoneauthentication/service/PhoneAuthenticationBanService.java
@@ -1,4 +1,4 @@
-package com.zoopi.client.service.phoneauthentication;
+package com.zoopi.client.phoneauthentication.service;
 
 import java.time.LocalDateTime;
