+++ module-model/src/main/java/com/zoopi/client/member/model/MemberAuthDto.java
@@ -1,4 +1,4 @@
-package com.zoopi.client.model.member;
+package com.zoopi.client.member.model;
 
 import static com.zoopi.util.Constants.*;
